pyproject.toml
[project]
name = "rsi-bot"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.5",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
    "python-telegram-bot>=22.4",
    "sqlalchemy>=2.0.43",
    "requests>=2.31.0",
    "numpy>=1.24.0",
    "websockets>=12.0",
    "aiohttp>=3.9.0",
]
